chore(cdr): Move httpmw to /coderd directory
httpmw is specific to coderd and should be scoped under coderd
