+++ coderd/coderd.go
@@ -12,7 +12,7 @@ import (
 	"cdr.dev/slog"
 	"github.com/coder/coder/database"
 	"github.com/coder/coder/httpapi"
-	"github.com/coder/coder/httpmw"
+	"github.com/coder/coder/coderd/httpmw"
 	"github.com/coder/coder/site"
 )
