+++ coderd/files.go
@@ -15,7 +15,7 @@ import (
 	"github.com/coder/coder/codersdk"
 	"github.com/coder/coder/database"
 	"github.com/coder/coder/httpapi"
-	"github.com/coder/coder/httpmw"
+	"github.com/coder/coder/coderd/httpmw"
 )
 
 func (api *api) postFile(rw http.ResponseWriter, r *http.Request) {
