+++ coderd/httpmw/apikey_test.go
@@ -16,7 +16,7 @@ import (
 	"github.com/coder/coder/database"
 	"github.com/coder/coder/database/databasefake"
 	"github.com/coder/coder/httpapi"
-	"github.com/coder/coder/httpmw"
+	"github.com/coder/coder/coderd/httpmw"
 )
 
 func randomAPIKeyParts() (id string, secret string) {
