+++ coderd/httpmw/userparam_test.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/coder/coder/database"
 	"github.com/coder/coder/database/databasefake"
-	"github.com/coder/coder/httpmw"
+	"github.com/coder/coder/coderd/httpmw"
 )
 
 func TestUserParam(t *testing.T) {
