+++ coderd/projectversions.go
@@ -12,7 +12,7 @@ import (
 	"github.com/coder/coder/codersdk"
 	"github.com/coder/coder/database"
 	"github.com/coder/coder/httpapi"
-	"github.com/coder/coder/httpmw"
+	"github.com/coder/coder/coderd/httpmw"
 )
 
 func (api *api) projectVersion(rw http.ResponseWriter, r *http.Request) {
