+++ coderd/workspaces.go
@@ -16,7 +16,7 @@ import (
 	"github.com/coder/coder/codersdk"
 	"github.com/coder/coder/database"
 	"github.com/coder/coder/httpapi"
-	"github.com/coder/coder/httpmw"
+	"github.com/coder/coder/coderd/httpmw"
 )
 
 func (api *api) workspace(rw http.ResponseWriter, r *http.Request) {
