+++ codersdk/workspaceresources.go
@@ -16,7 +16,7 @@ import (
 	"nhooyr.io/websocket"
 
 	"github.com/coder/coder/database"
-	"github.com/coder/coder/httpmw"
+	"github.com/coder/coder/coderd/httpmw"
 	"github.com/coder/coder/peer"
 	"github.com/coder/coder/peerbroker"
 	"github.com/coder/coder/peerbroker/proto"
